+++ src/store/extra-argument.ts
@@ -1,0 +1,3 @@
+export interface ExtraArgument {
+  api: string;
+}
